paillier: offer EncryptWithNonce and EncryptAndNonce in API
This commit adds additional encryption features that offer
more precise control over the chosen nonces.
EncryptWithNonce will compute the CRT optimized Paillier
encryption with a provided nonce. EncryptAndNonce will
generate a random nonce for encryption, but return it to
the caller. Together, this provides more flexibility in
using the Paillier operations for other protocols that
rely on Paillier as a primitive.